Tagid-RF/be/requirements.txt
fastapi==0.112.0
uvicorn==0.34.0
prisma==0.15.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pydantic==2.11.3
pydantic-settings==2.8.1
python-json-logger==3.3.0
email-validator==2.2.0
python-dotenv==1.1.0
# Security packages
cryptography==44.0.2
# Uncomment for rate limiting
# slowapi==0.1.8
# redis==5.0.1
# M-200 uses custom protocol implementation
# Google OAuth dependencies
google-auth==2.38.0
google-api-python-client==2.166.0
firebase-admin==6.5.0
stripe==11.3.0

sqlalchemy==2.0.36
psycopg[binary]
psycopg2-binary==2.9.11
alembic==1.12.1
websockets==12.0
python-dateutil==2.8.2
pyserial==3.5  # For serial communication with RFID reader
qrcode[pil]==7.4.2  # For generating product QR codes
bcrypt==4.2.0  # For password hashing
# M-200 Gate Reader uses native protocol implementation (app/services/m200_protocol.py)

# Testing dependencies
pytest==8.3.2
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.0
faker==36.2.2
